requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
aiohttp
numpy
pydantic
beautifulsoup4
html2text
tqdm
markdown
python-multipart
setuptools
jinja2
aiosqlite